requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
Automat==22.10.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==6.0.0
blobfile==2.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
constantly==15.1.0
contourpy==1.0.7
cryptography==39.0.0
cssselect==1.2.0
cycler==0.11.0
dataclasses-json==0.5.7
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.9.0
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.30
gpt-index==0.4.1
greenlet==2.0.2
huggingface-hub==0.12.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipykernel==6.20.2
ipython==8.8.0
ipython-genutils==0.2.0
isoduration==20.11.0
itemadapter==0.7.0
itemloaders==1.0.6
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-server-mathjax==0.2.6
jupyter_client==7.4.9
jupyter_core==5.1.5
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab==3.5.3
jupyterlab-git==0.41.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.19.0
kiwisolver==1.4.4
langchain==0.0.83
lxml==4.9.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
mistune==2.0.4
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.9
nbdime==3.1.1
nbformat==5.7.3
nest-asyncio==1.5.6
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
openai==0.26.5
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parsel==1.7.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
plotly==5.13.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
Protego==0.2.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.17
pydantic==1.10.4
PyDispatcher==2.0.6
Pygments==2.14.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.1
python-json-logger==2.0.4
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
queuelib==1.6.2
regex==2022.10.31
requests==2.28.2
requests-file==1.5.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-learn==1.2.1
scipy==1.10.0
Scrapy==2.7.1
seaborn==0.12.2
Send2Trash==1.8.0
service-identity==21.1.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
stack-data==0.6.2
tenacity==8.1.0
terminado==0.17.1
threadpoolctl==3.1.0
tiktoken==0.2.0
tinycss2==1.2.1
tldextract==3.4.0
tokenizers==0.13.2
tomli==2.0.1
torch==1.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
transformers==4.26.1
Twisted==22.10.0
typing-inspect==0.8.0
typing_extensions==4.4.0
uri-template==1.2.0
urllib3==1.26.14
w3lib==2.1.1
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
yarl==1.8.2
zope.interface==5.5.2
